Add use_classic_font_in_k5 account setting
Adds a setting to the account settings page that is visible when the
k5_font_selection flag is enabled and Canvas for Elementary is checked.
This lets users pick the font to use in k5 - balsamiq (default) or
lato (like in classic Canvas). This setting is inherited to child
accounts and, like the k5 setting, once set, is locked for child
accounts.

When the setting is mutated in the UI, the account ID is added
or removed from the root account's list of accounts where lato should
be used (also like the k5 setting). This will let us determine at
runtime which font should be shown to the user in account-agnostic
spaces, like the dashboard.

Unlike the k5 setting, when a parent account modifies its
chosen font settings (or disables k5 entirely), we need to remove
font settings from the descendent accounts so that if the parent
choses the k5 font, a child's potential classic font selection doesn't
conflict.

closes LS-3828
flag = k5_font_selection

Test plan:
 - On an account, enable k5 and change the font setting
 - Expect the setting to persist on page reload
 - Expect the setting to be locked on subaccounts (and reflect the
   chosen font)
 - Expect root_account.settings[:k5_classic_font_accounts] to contain
   the list of account ids where classic font has been selected

Change-Id: Ie66e6127d184ebe254f431715bd0ca375ca46418
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/312083
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>