Fix bug where sso-icon does not populate
User login sso-icons were not showing in login information of the
user page. This was due to the lookup code for the svg not working,
when the view is checking if the svg exists. In my testing, the old
lookup code works for erb files but not svg files. This code change
will correctly check if the svg file exists.

fixes FOO-3368
flag=none

test plan:
 - Import user with authentication_provider_id set to canvas
 - Confirm login information of user shows canvas icon on user page

Change-Id: I09736471b5c317ce1b73a5f44439095c5547de2a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/312704
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Ajmal Esa <ajmal.esa@instructure.com>
Product-Review: Ajmal Esa <ajmal.esa@instructure.com>