peer review comments under recent feedback
fix for peer review comments not showing up under recent feedback if a
teacher had first not graded or commented on a submission. this was due
to a check on the posted_at date which is only set when a teacher does
an action on a submission. now we check if the submission is posted OR
if the teacher has set manual posting to false

fixes EVAL-465
flag=none

test plan:
- as teacher, create a peer review assignment
  - set manual posting to false
  - assign two students (A,B) to review each other
- as student A, submit for the assignment
- as student B, submit for the assignment
- as student A, review student B's submission
- as student B, review student A's submission
- as student B, verify that under the Home page > Recent Feedback
  section, you see peer review assignment listed

Change-Id: I1f4496462942fd8b21e765baaef21a4ac5a7b529
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/313094
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>