Respect content can_publish? or can_unpublish? methods
This is similar to the ContextModulesHelper that verifies if a module
item can be published or unpublished on the frontend. The bulk update
all happens in the backend so those checks weren't preventing calls for
those items. Now when publishing or unpublishing we check if the content
has that method and if it does we will respect the output of that before
making a change to the module item.

fixes LS-3928
flag=module_publish_menu

test plan:
- Have a module with many types of module items including one attachment
- Publish all the items and refresh the page to make sure everything is
published.
- Use the bulk menu to unpublish all items
- Verify the attachment was not unpublished. All attachments return false
for can_unpublish? to require the use of the file modal.
- Verify all the other items that were possible to unpublish were
unpublished.
- You can also use an assignment that is published and has a student
submission to verify things don't unpublish since assignments with
submissions cannot be unpublished

Change-Id: I7d290de84942b53c840711d95c70c74e6533d191
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/315669
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>