spec: Look for newly-added routes for dev key scopes
We need to have an up-to-date list of api routes
hard-coded so that we will notice when one changes. This
test will fail if a route is defined in routes.rb but
not listed in the last_known_scopes.yml list.

Change-Id: I641db1a77859b021dcbcb4e81ed20310e953972f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/317314
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>