Resolve a race condition in enrollment imports
In the specific case of a diff mode import where the enrollment role
is being changes, it's possible for a parallel import to run deleting
the old enrollment before the new one is created. This results in a
race condition where the group memberships for a user are deleted.

This changeset adds a check to the enrollment batch state updater to
ensure that the enrollment is not active elsewhere in the diffed import
before deleting a group membership.

Fixes FOO-3448
flag=none

Test Plan:
 * Enable parallel imports (SIS plugin config)
 * Create a custom student role
 * Run a SIS import with enrollments for a course (with diff mode)
 * Assign students to groups in the course
 * Run a SIS import that changes role for the same students under the
   same diff mode id.
 * Verify that the students are still in the groups they were assigned
   to in the previous import.

Change-Id: Id71427b04de457134ba52d8b6d3f8a58b2945806
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/319487
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jason Perry <jason.perry@instructure.com>
Product-Review: Jason Perry <jason.perry@instructure.com>