Keep track of quoted_entry_id when creating new entries
Since we are going to allow threads to be nested up to 3 levels deep
We need a more precise way to keep track of the quoted entry id since
the quoted entry will not always be the parent_id.

This will allow us to track the id of the quoted entry and display it.

refs: VICE-3243
flag=split_screen_view

Test Plan:
1. Quote replies on discussion redesign
2. Verify that quoted_entry_id is the id of the entry that you
started the quote on.

Note that for 3rd and 4th level repleis for splitscreen/inline view that the
Actual displayed quoted entry after the creation will not match.
This is a bug that will be fixed in VICE-3225

Change-Id: I30158f8df7aa50033336830f5d245616bfc655df
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/319692
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
Product-Review: Jason Gillett <jason.gillett@instructure.com>