enhanced gradebook lock down query param routing
currently the new enhanced individual gradebook is able to be accessed
with the FF off if the user knows to navigate to the gradebook url with
the query param ?version=individual_enhanced. This commit adds a check
on the backend during the routing that will now direct the user to the
traditional gradebook grid if the FF is off.

closes EVAL-3419

flag=individual_gradebook_enhancements

test plan:
- navigate to /courses/<course_id>/gradebook?version=individual_enhanced
  and verify that you are directed to traditional gradebook

Change-Id: Ief0c905ffd919a0ea87a3c7643a07fff3ab9ac67
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/325291
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Rohan Chugh <rohan.chugh@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>