Upgrade Canvas to InstUI version 8 (second try)
Refs FOO-3190
Refs FOO-3708
Refs FOO-3913
flag=none
[fsc-max-nodes=30] [fsc-timeout=60]

Uprevs InstUI and addresses all post-uprev tasks. Also now that
InstUI is fully TypeScript, a bunch of type errors appeared
which needed to be addressed.

It builds against the commit of MRA that will have to merge
concurrently with this.

Still to do: theme -> themeOverride which is a separate commit
off of this one that is ready to go.

Test plan:
* cross your fingers
* (this will need some extensive review before merging)

Change-Id: I45a1138530d415398b019cfedcf8a3b57a051912
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/329443
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Norton <rnorton@instructure.com>
Reviewed-by: Isaac Moore <isaac.moore@instructure.com>
Product-Review: Ryan Norton <rnorton@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>