properly check for admin access
Teachers with admin account access in a sub account was wrongfully
getting access to assignment/quiz setting that should not be available
to edit when the grading period is closed. This was due to us just
checking if they had an admin role. We were not checking if the admin
role is within the account the course is in. This ticket passes through
a new js_env variable to rely on for checking admin access.

closes EVAL-3547
flag=none

Test Plan:
  - Prerequisites: Have two sub-accounts, and a course in one sub-account
  that has a term with grading periods. Add an assignment, quiz, and new
   quiz where the grading period is closed. This will require
   submissions to exist on each.
  - Add a user as an admin to sub-account 1
  - Enroll the same user as a teacher into the course in sub-account 2
  - As the user, edit the assignment/quiz created in the prerequisite
  - Notice that a message about it being in a closed grading period
  appears
  - Notice that many of the assignment fields are disabled and the ones
  that are enabled can be edited and saved successfully
  - As the admin, edit the assignment/quiz created in the prerequisite
  - Ensure all settings are enabled and available to be edited by the
  the admin

Change-Id: I1649a4429c958cc85fcbda9f1e089408850c1924
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/331559
Reviewed-by: Derek Williams <derek.williams@instructure.com>
QA-Review: Derek Williams <derek.williams@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>