fix isMismatch bug in enrollment tree
The isMismatch was not being set (on load) for parent course
(in UI) when a course had child sections checked and the chosen
role did not match the checked enrollment. This caused the mismatch
icon to not display in the UI for the parent course. This commit
fixes that bug.

closes FOO-4115
flag=temporary_enrollments

test plan:
- enable temporary_enrollments feature
- enable Temporary Enrollments permissions
- create and publish a course
- add a teacher to course
- create sections for a course
- add a teacher to a section
- go to course/people page
- create temporary enrollments
- navigate to last screen of modal
- choose a role that sets a mismatch for section
- expect parent course to display mismatch icon

Change-Id: I0ed6126d93c73fca24a14e08b380b7901de2d0f7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/334145
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Product-Review: Michael Hulse <michael.hulse@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>