be more exact when speedgrader considers concluded
Speedgrader in the all sections case will now only consider a student
concluded if they are concluded in all sections they are enrolled in.
When in section specific mode, and with concluded enrollments on, it
will consider a student as concluded when viewing that section, but not
in other sections where the student may be active.

fixes EVAL-3656
flag = none

test plan:
 - create a course with multiple sections
 - create a student and place them in two sections
 - conclude the student in one section
 - go to speedgrader and view the student in the section they are not
   concluded in, and verify that they are not marked as concluded
 - go to speedgrader and view the student in the section they are
   concluded in, and verify that they are marked as concluded

Change-Id: I9c9fde7295fbabd2eb968df06fe15c8254e572b5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/334571
QA-Review: Samuel Lee <samuel.lee@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Derek Williams <derek.williams@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>