Move dynamic reg access token to url parameter
why

The specification only dictates that the access token be sent in the
Authorization header for the registration request. Before this change,
it was also required for the openid-configuration url, too, which is
wrong.

test plan:

register an lti tool with the dynamic registration flow. Note that the
openid configuration url includes the access token as a query parameter.

flag=lti_dynamic_registration

Change-Id: I36bc61c6e503e5b42fbd40cead0af131b596809b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/338642
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>