spec: Move page_focused set up outside of test module
This is to fix some of the failing tests when upgrading to
jquery 1.9. The fix is to move the set up outside of the
test module because setting up the event listeners on
the window between each test causes failures in
jQuery 1.9.

refs FOO-4258
flag=none

Test Plan:
- Jenkins passes

[skip-crystalball]

Change-Id: If8b9fce0e4a5bf38f86a0ed65cba4aa137afa3b9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/340800
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Angela Gomba <angela.gomba@instructure.com>
Product-Review: Angela Gomba <angela.gomba@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>