Make DatetimeField loading more bulletproof
Refs FOO-2801
flag=none

Dynamic loading of locale JSONs leaves open the possibility
that a locale might load that is mis-configured, missing
some of our standard translation keys. This can create
seemingly-unrelated breakage.

In particular, DatetimeField (which is how <input> elements
magically become jQuery datepickers) was not populating its
own data structures in a very resilient way. It assumed that
the I18n strings for things like month names and weekday
names are "of course" present in the loaded locale. But if
for some reason a key was missing, DatetimeField's startup
would throw an error, resulting in a blank UI page at worst
and at least malfunctioning datepickers at best.

This wraps the critical I18n lookups in a try/catch, so that
a fallback to US English can be loaded as a last resort if
nothing is available in the selected user's language.

Test plan:
* Datepickers in DatetimeFields should load and work as
  normal
* Contrive to temporarily break a locale file by removing one
  of the important keys (such as date.month_names), and
  restart your instance
* Go to your user settings page (/profile/settings) with your
  JavaScript console open and set your own language to the one
  whose locale file you have just corrupted
* Reload; the settings page should load normally
* The console should print a warning saying that US English
  will be used for the datepicker, but the page should otherwise
  come up and function normally

Change-Id: I8b38eaa8983a4500a144781672dccdfeafd942e6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/342075
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jason Perry <jason.perry@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>