remove dead code breaking in ruby 3.2
refs AE-855

This code never worked correctly as the =~ operation on an object always returns nil, and it has been removed in Ruby 3.2.

Based off of the regex, it appears that the original intention was to use `input[:recipients]` as one possible input from the specs is `course_12_students`.

Remove the code for now as it has never worked, and allow the owning team to restore it later to reduce risk.

Change-Id: I5813b9badf0aa1c0a8f0164a8d71d7465ae24870
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/343973
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Aaron Ogata <aogata@instructure.com>
Product-Review: Aaron Ogata <aogata@instructure.com>